+++ src/js/tutorial/tutorial.module.js
@@ -1,0 +1,9 @@
+import TutorialCtrl from "./tutorialCtrl";
+
+(() => {
+  "use strict";
+
+  angular.module("modwatchuploader.tutorial", [])
+    .controller("TutorialCtrl", TutorialCtrl);
+
+})();