+++ src/js/tutorial/tutorialCtrl.js
@@ -1,0 +1,7 @@
+LoginCtrl.$inject = ["APIService"];
+
+export default LoginCtrl;
+
+function LoginCtrl(APIService) {
+
+}