+++ src/js/upload/upload.module.js
@@ -1,0 +1,9 @@
+import UploadCtrl from "./uploadCtrl";
+
+(() => {
+  "use strict";
+
+  angular.module("modwatchuploader.upload", [])
+    .controller("UploadCtrl", UploadCtrl);
+
+})();